+++ command/team_test.go
@@ -9,3 +9,7 @@ import (
 func TestAddTeamCommand_implement(t *testing.T) {
 	var _ cli.Command = &AddTeamCommand{}
 }
+
+func TestDeleteTeamCommand_implement(t *testing.T) {
+	var _ cli.Command = &DeleteTeamCommand{}
+}
